fix(cli): bound the connect and upgrade, and cancel a query on revocation
Six review findings from PR #1085, each verified against the code first.

The two UNBOUNDED steps in an otherwise fully budgeted function: connect and
the WebSocket upgrade. A process that accepts on a PORT_RANGE port and never
completes the upgrade -- a wedged previous instance whose listener is still
bound but whose accept loop stopped -- parked the whole port loop forever. Both
are now wrapped, and the whole invocation carries an outer deadline. Proven by
standing up a real TCP squatter: the built binary now returns app_not_running
in ~15s instead of never returning.

The UNC guard only rejected a literal \ or // prefix, but Windows treats the
separators interchangeably, so /\host\share and \/host/share were absolute,
parsed as a UNC root, and passed -- defeating the SMB/NTLM mitigation the doc
claimed. Now checks either separator in either position.

The unknown-flag branch still echoed argv into the exit-2 detail; that leak was
fixed for the unknown-verb branch and missed one case over, with the test
pinning only the branch that was fixed. Usage errors now also carry a null
resource so every exit-2 reply has the same shape.

An in-flight agent.query was not cancelled on token revocation -- a gap
introduced by the earlier fix that moved dispatch off the read loop, which is
where fresh defects come from. It now races a per-connection cancellation token
cancelled at the shared teardown site. Two honest limits: best-matches' blocking
CPU pass has no cooperative checkpoint, so the reply is suppressed and the
connection freed but running compute is not preempted; and spawn_answer_assist
has the SAME stale-reply-after-cancellation gap, which was left unfixed rather
than copied, and needs its own follow-up.

Accumulated docs pushed agent_cli.rs past R8's 1400-line cap, so the test module
moved to agent_cli/tests.rs, the split convention commands/autopilot.rs uses.
